REFACTOR: SwipeCarousel을 공용 레이아웃으로 분리
